fix validation logic for dates in duckdb and postgres
This fix is pre-requisite to using test cases from substrait directly.

substrait testcases have examples which can only contain the date part
but the python client returns a datetime object.

Converting these to strings and comparing is not quite right.
e.g., '2024-03-01' == '2024-03-01:00:00:00' will fail if we do string compare.

The fix is to not compare the empty fields hrs, mins, secs, etc if we only
contain the date part